requirements.txt
pillow~=11.3.0
stegano~=2.0.0